requirements.txt
# Airflow packages ---------------
apache-airflow==2.7.0 # Apache Airflow package
apache-airflow-providers-cncf-kubernetes==6.0.0  # Airflow provider, if Kubernetes is used
#---------------------------------

# OCR and image processing -------
pytesseract # For OCR text extraction (Interface to OCR engine)
tesseract==0.1.3 # For OCR text extraction (OCR engine)
Pillow==9.2.0 # For image processing
PyMuPDF # High-performance Python library for data extraction, analysis, conversion & manipulation of PDF
#---------------------------------

# OpenAI --------
openai==1.52.2 # For OpenAI API interaction
#----------------

# Computer vision tools ----------
opencv-python-headless # Real-time optimized Computer Vision library, tools
transformers # Provides APIs and tools to easily download and train pretrained models
timm # Library containing SOTA computer vision models, layers, utilities, optimizers, schedulers, data-loaders, augmentations, and training/evaluation scripts
#---------------------------------

# Web App framework -----
streamlit # Python framework that deliver dynamic data apps
#------------------------

# String similarity and Natural Language tools --------
Levenshtein # Extension module contains functions for fast computation of Levenshtein distance
nltk # Library for symbolic and statistical natural language processing
#------------------------------------------------------

# Other utilities ----------------
numpy==1.26.4 # Library that large, multi-dimensional arrays and matrices
pandas==1.5.0 # For working with DataFrames
PyYAML==6.0 # For handling YAML config files
openpyxl # Library to read/write Excel 2010 xlsx/xlsm/xltx/xltm files
xlsxwriter # Library for writing files in the Excel 2007+ XLSX file format
requests==2.28.1 # For handling HTTP requests
scikit-learn # Library for machine learning
tabulate # Library that provides a command-line utility to help you display tables directly from the command line or terminal
#---------------------------------